+++ src/routing/http_wrapper.cpp
@@ -22,11 +22,13 @@ const std::string HttpWrapper::HTTPS_PORT = "443";
 HttpWrapper::HttpWrapper(const std::string& profile,
                          const Server& server,
                          const std::string& matrix_service,
+                         const std::string& matrix_durations_key,
                          const std::string& route_service,
                          const std::string& extra_args)
   : Wrapper(profile),
     _server(server),
     _matrix_service(matrix_service),
+    _matrix_durations_key(matrix_durations_key),
     _route_service(route_service),
     _extra_args(extra_args) {
 }
@@ -72,9 +74,9 @@ std::string HttpWrapper::send_then_receive(const std::string& query) const {
   auto end = response.rfind("}");
   assert(end != std::string::npos);
 
-  std::string json_content = response.substr(start, end - start + 1);
+  std::string json_string = response.substr(start, end - start + 1);
 
-  return json_content;
+  return json_string;
 }
 
 std::string HttpWrapper::ssl_send_then_receive(const std::string& query) const {
@@ -121,9 +123,9 @@ std::string HttpWrapper::ssl_send_then_receive(const std::string& query) const {
   auto end = response.rfind("}");
   assert(end != std::string::npos);
 
-  std::string json_content = response.substr(start, end - start + 1);
+  std::string json_string = response.substr(start, end - start + 1);
 
-  return json_content;
+  return json_string;
 }
 
 std::string HttpWrapper::run_query(const std::string& query) const {
@@ -131,17 +133,30 @@ std::string HttpWrapper::run_query(const std::string& query) const {
                                       : send_then_receive(query);
 }
 
+void HttpWrapper::parse_response(rapidjson::Document& json_result,
+                                 const std::string& json_content) const {
+#ifdef NDEBUG
+  json_result.Parse(json_content.c_str());
+#else
+  assert(!json_result.Parse(json_content.c_str()).HasParseError());
+#endif
+}
+
 Matrix<Cost> HttpWrapper::get_matrix(const std::vector<Location>& locs) const {
   std::string query = this->build_query(locs, _matrix_service);
-  std::string json_content = this->run_query(query);
+  std::string json_string = this->run_query(query);
 
   // Expected matrix size.
   std::size_t m_size = locs.size();
 
-  rapidjson::Document infos;
-  this->parse_response(infos, json_content);
+  rapidjson::Document json_result;
+  this->parse_response(json_result, json_string);
+  this->check_response(json_result, _matrix_service);
 
-  assert(infos["durations"].Size() == m_size);
+  if (!json_result.HasMember(_matrix_durations_key.c_str())) {
+    throw Exception(ERROR::ROUTING, "Missing " + _matrix_durations_key + ".");
+  }
+  assert(json_result[_matrix_durations_key.c_str()].Size() == m_size);
 
   // Build matrix while checking for unfound routes ('null' values) to
   // avoid unexpected behavior.
@@ -150,19 +165,18 @@ Matrix<Cost> HttpWrapper::get_matrix(const std::vector<Location>& locs) const {
   std::vector<unsigned> nb_unfound_from_loc(m_size, 0);
   std::vector<unsigned> nb_unfound_to_loc(m_size, 0);
 
-  for (rapidjson::SizeType i = 0; i < infos["durations"].Size(); ++i) {
-    const auto& line = infos["durations"][i];
+  for (rapidjson::SizeType i = 0; i < m_size; ++i) {
+    const auto& line = json_result[_matrix_durations_key.c_str()][i];
     assert(line.Size() == m_size);
     for (rapidjson::SizeType j = 0; j < line.Size(); ++j) {
-      if (line[j].IsNull()) {
+      if (duration_value_is_null(line[j])) {
         // No route found between i and j. Just storing info as we
         // don't know yet which location is responsible between i
         // and j.
         ++nb_unfound_from_loc[i];
         ++nb_unfound_to_loc[j];
       } else {
-        auto cost = round_cost(line[j].GetDouble());
-        m[i][j] = cost;
+        m[i][j] = get_duration_value(line[j]);
       }
     }
   }
@@ -193,16 +207,17 @@ void HttpWrapper::add_route_info(Route& route) const {
   std::string query =
     build_query(non_break_locations, _route_service, _extra_args);
 
-  std::string json_content = this->run_query(query);
+  std::string json_string = this->run_query(query);
 
-  rapidjson::Document infos;
-  parse_response(infos, json_content);
+  rapidjson::Document json_result;
+  parse_response(json_result, json_string);
+  this->check_response(json_result, _route_service);
 
   // Total distance and route geometry.
-  route.distance = round_cost(get_total_distance(infos["routes"][0]));
-  route.geometry = std::move(infos["routes"][0]["geometry"].GetString());
+  route.distance = round_cost(get_total_distance(json_result));
+  route.geometry = get_geometry(json_result);
 
-  auto nb_legs = get_legs_number(infos["routes"][0]);
+  auto nb_legs = get_legs_number(json_result);
   assert(nb_legs == non_break_locations.size() - 1);
 
   double sum_distance = 0;
@@ -226,7 +241,7 @@ void HttpWrapper::add_route_info(Route& route) const {
     // distance after current route leg.
     auto& next_step = route.steps[steps_rank + number_breaks_after[i] + 1];
     Duration next_duration = next_step.duration - step.duration;
-    double next_distance = get_distance_for_leg(infos["routes"][0], i);
+    double next_distance = get_distance_for_leg(json_result, i);
 
     // Pro rata temporis distance update for breaks between current
     // non-breaks steps.