+++ src/routing/osrm_routed_wrapper.h
@@ -21,17 +21,25 @@ private:
                                   const std::string& service,
                                   const std::string& extra_args) const override;
 
-  virtual void parse_response(rapidjson::Document& input,
-                              const std::string& json_content) const override;
+  virtual void check_response(const rapidjson::Document& json_result,
+                              const std::string& service) const override;
+
+  virtual bool
+  duration_value_is_null(const rapidjson::Value& matrix_entry) const override;
+
+  virtual Cost
+  get_duration_value(const rapidjson::Value& matrix_entry) const override;
 
   virtual double
-  get_total_distance(const rapidjson::Value& route) const override;
+  get_total_distance(const rapidjson::Value& result) const override;
 
   virtual unsigned
-  get_legs_number(const rapidjson::Value& route) const override;
+  get_legs_number(const rapidjson::Value& result) const override;
 
-  virtual double get_distance_for_leg(const rapidjson::Value& route,
+  virtual double get_distance_for_leg(const rapidjson::Value& result,
                                       rapidjson::SizeType i) const override;
+
+  virtual std::string get_geometry(rapidjson::Value& result) const override;
 
 public:
   OsrmRoutedWrapper(const std::string& profile, const Server& server);